pyproject.toml
[tool.poetry]
name = "mobile_test"
version = "1.0"
description = "mobile testing using browserstack"
authors = ["andrechizh8 <andrechizh.ru@yandex.ru>"]


[tool.poetry.dependencies]
python = "^3.10"
selene = {version = "^2.0.0-beta.8", allow-prereleases = true}
Appium-Python-Client = "^2.6.1"
pytest = "^7.1.3"
pytest-xdist = "^2.5.0"
allure-pytest = "^2.10.0"
requests = "^2.28.1"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"


[tool.pytest.ini_options]
addopts = " --alluredir=allure-results "